Broke costing envelope, lighting, ventilation, and heating and cooling costing into their own files.  Moved lighting costing method into lighting_costing.rb and envelope and constructions costing into envelope_costing.rb.
