+++ measures/btap_results/resources/heating_cooling_costing.rb
@@ -1,0 +1,5 @@
+class BTAPCosting
+  def heting_costing
+    puts "test heating"
+  end
+end
