+++ measures/btap_results/resources/lighting_costing.rb
@@ -1,0 +1,5 @@
+class BTAPCosting
+  def cost_audit_lighting(model)
+    puts "test lighting"
+  end
+end
