+++ measures/btap_results/resources/ventilation_costing.rb
@@ -1,0 +1,18 @@
+class BTAPCosting
+  def ahu_costing(model, prototype_creator)
+    air_loops = model.getAirLoopHVACs
+
+    model.getAirLoopHVACs.each do |airloop|
+      airloop.thermalZones.each do |tz|
+        tz_name = tz.nameString
+
+        tz.spaces.each do |space|
+          sys_info = prototype_creator.get_sys_selection_and_loads(model, space)
+
+          puts "hello"
+        end
+      end
+    end
+    puts "hello"
+  end
+end
